+++ ContentPatcher/Framework/ConfigModels/ContentConfig.cs
@@ -25,7 +25,7 @@ namespace ContentPatcher.Framework.ConfigModels
         public CustomLocationConfig?[] CustomLocations { get; }
 
         /// <summary>The changes to make.</summary>
-        public PatchConfig[] Changes { get; }
+        public PatchConfig[] Changes { get; private set; }
 
         /// <summary>The schema for the <c>config.json</c> file (if any).</summary>
         public InvariantDictionary<ConfigSchemaFieldConfig?> ConfigSchema { get; }
@@ -49,6 +49,14 @@ namespace ContentPatcher.Framework.ConfigModels
             this.CustomLocations = customLocations ?? Array.Empty<CustomLocationConfig>();
             this.Changes = changes?.WhereNotNull().ToArray() ?? Array.Empty<PatchConfig>();
             this.ConfigSchema = configSchema ?? new InvariantDictionary<ConfigSchemaFieldConfig?>();
+        }
+
+        /// <summary>Replace the list of patches in this content pack.</summary>
+        /// <param name="changes">The changes to apply instead.</param>
+        [Obsolete("This is only intended to support specific cases in the low-level migration code. This shouldn't be used by most code.")]
+        public void SetChanges(PatchConfig[] changes)
+        {
+            this.Changes = changes;
         }
     }
 }