+++ ContentPatcher/Framework/Migrations/Migration_1_16.cs
@@ -17,12 +17,12 @@ namespace ContentPatcher.Framework.Migrations
             : base(new SemanticVersion(1, 16, 0)) { }
 
         /// <inheritdoc />
-        public override bool TryMigrate(ContentConfig content, [NotNullWhen(false)] out string? error)
+        public override bool TryMigrate(ref PatchConfig[] patches, [NotNullWhen(false)] out string? error)
         {
-            if (!base.TryMigrate(content, out error))
+            if (!base.TryMigrate(ref patches, out error))
                 return false;
 
-            foreach (PatchConfig patch in content.Changes)
+            foreach (PatchConfig patch in patches)
             {
                 // 1.16 adds Include
                 if (this.HasAction(patch, PatchType.Include))