+++ ContentPatcher/docs/release-notes.md
@@ -13,6 +13,9 @@ When releasing a format change, don't forget to update the smapi.io/json schema!
 * Improved error when an `EditData` patch can't convert the data to the asset type.
 * Fixed setting map/tile properties to null no longer removing them.
 
+## Upcoming release
+* Fixed some migrations not applied to files loaded via `Action: Include`.
+
 ## 1.29.1
 Released 31 March 2023 for SMAPI 3.18.1 or later.